Refactored CameraPermissions into an Interface and AndroidCameraPermissions into the implementation to make it easier to fake a CameraPermissions in tests.
